requirements.txt
# Core scientific packages
numpy
h5py
pandas[parquet]

# Visualization
matplotlib

# Geospatial
shapely
fiona
geopandas
pyproj

# Testing
hypothesis[numpy]
pytest
pytest-cov
pytest-repeat

# Project dependencies
numba
qcore @ git+https://github.com/ucgmsim/qcore.git
pyyaml
tqdm
typer

# Tool dependencies
cartopy
pytz
requests
